Create AddOrEditObject.sql
This procedure checks if the entered id already exists, if it does then the object in the database will be updated with the new values for column1, column2 and column3. If the id does not exist then a new object will be created with the entered id and values for the other columns.